move data context to core
