+++ Anlab.Core/Models/OrderDetails.cs
@@ -1,13 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
 using Anlab.Core.Domain;
-using Anlab.Core.Models;
 
-
-namespace AnlabMvc.Models.Order
+namespace Anlab.Core.Models
 {
     public class OrderDetails
     {
